% *=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=* 
% ** Copyright UCAR (c) 1992 - 2004 
% ** University Corporation for Atmospheric Research(UCAR) 
% ** National Center for Atmospheric Research(NCAR) 
% ** Research Applications Program(RAP) 
% ** P.O.Box 3000, Boulder, Colorado, 80307-3000, USA 
% ** 2004/9/1 14:13:29 
% *=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=* 
\name{analysis.dat}
\alias{analysis.dat}
\title{
Spatial Observation Dataset.
}
\description{
This spatial data set is used by the \code{\link{int.scale.verify}}
function.  This consists of the outcome of the weather forecast presented
in \code{\link{forecast.dat}}. This data was provided by Barabara Casati and is
from the Met Office.
}
\keyword{datasets}
% docclass is data
% Converted by Sd2Rd version 1.21.
