% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Imprecision_fct_boxesPrec.R
\name{boxesPrec}
\alias{boxesPrec}
\title{boxesPrec}
\usage{
boxesPrec(
  model,
  x_var,
  colours,
  qc_level,
  test_claims,
  claims_data,
  ci_interval
)
}
\arguments{
\item{model}{rstanarm model object}

\item{x_var}{character denoting column containing x values}

\item{colours}{character vector of colours}

\item{qc_level}{numeric or character denoting QC level}

\item{test_claims}{boolean denoting whether to test against claims}

\item{claims_data}{numeric vector of manufacturer's claims}

\item{ci_interval}{numeric CI interval}
}
\value{
list of shiny fluidRows -> boxes -> valueBoxes
}
\description{
boxesPrec
}
