% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Reference_fct_boxesRef.R
\name{boxesRef}
\alias{boxesRef}
\title{boxesRef}
\usage{
boxesRef(model, prior_location, ci_interval, model_type, bayes_factor)
}
\arguments{
\item{model}{rstanarm model object}

\item{prior_location}{numeric prior location}

\item{ci_interval}{numeric CI interval}

\item{model_type}{character string of: "reference" or "reference_varying"}

\item{bayes_factor}{bayesfactor object returned from getBayesFactor()}
}
\value{
list of shiny valueBoxes
}
\description{
boxesRef
}
