% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Comparison_fct_calcBAStats.R
\name{calcBAStats}
\alias{calcBAStats}
\title{calcBAStats}
\usage{
calcBAStats(data, value_x1, value_y1, value_x2 = NULL, value_y2 = NULL, method)
}
\arguments{
\item{data}{data.frame input}

\item{value_x1}{character string denoting column containing method 1 values}

\item{value_y1}{character string denoting column containing method 2 values}

\item{value_x2}{character string denoting column containing method 1 duplicate values}

\item{value_y2}{character string denoting column containing method 2 duplicate values}

\item{method}{character string denoting difference type}
}
\value{
HTML text summarising analysis
}
\description{
calcBAStats
}
