% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calcCompCoef.R
\name{calcCompCoef}
\alias{calcCompCoef}
\title{calcCompCoef}
\usage{
calcCompCoef(data, method, model)
}
\arguments{
\item{data}{data.frame input}

\item{method}{character string denoting analysis method type}

\item{model}{mcr or brms object}
}
\value{
HTML text summarising model coefficients
}
\description{
calcCompCoef
}
