% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calcCompTest.R
\name{calcCompTest}
\alias{calcCompTest}
\title{calcCompTest}
\usage{
calcCompTest(
  data,
  model,
  value_x1,
  value_y1,
  value_x2 = NULL,
  value_y2 = NULL,
  method,
  ci_interval
)
}
\arguments{
\item{data}{data.frame input}

\item{model}{mcr or brms model object}

\item{value_x1}{character string denoting column containing method 1 values}

\item{value_y1}{character string denoting column containing method 2 values}

\item{value_x2}{character string denoting column containing method 1 duplicate values}

\item{value_y2}{character string denoting column containing method 2 duplicate values}

\item{method}{character string denoting test method}

\item{ci_interval}{numeric CI interval}
}
\value{
list of shiny valueBoxes
}
\description{
calcCompTest
}
