% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Diagnostic_fct_calcDiagCoef.R
\name{calcDiagCoef}
\alias{calcDiagCoef}
\title{calcDiagCoef}
\usage{
calcDiagCoef(data, model, col_value, positive, ci_interval)
}
\arguments{
\item{data}{data.frame input}

\item{model}{mcr or brms model object}

\item{col_value}{character string representing column of values}

\item{positive}{character string representing a "positive" case}

\item{ci_interval}{numeric CI interval}
}
\value{
list of shiny valueBoxes
}
\description{
calcDiagCoef
}
