% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checkInputFile.R
\name{checkInputFile}
\alias{checkInputFile}
\title{checkInputFile}
\usage{
checkInputFile(input_env, input_id)
}
\arguments{
\item{input_env}{shiny input object}

\item{input_id}{character string for inputId}
}
\value{
shinyFeedback object
}
\description{
checkInputFile
}
