% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checkInputLevels.R
\name{checkInputLevels}
\alias{checkInputLevels}
\title{checkInputLevels}
\usage{
checkInputLevels(input_env, input_id, input_data)
}
\arguments{
\item{input_env}{shiny input object}

\item{input_id}{character string for inputId}

\item{input_data}{reactive shiny object representing data}
}
\value{
shinyFeedback object
}
\description{
checkInputLevels
}
