% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checkInputNull.R
\name{checkInputNull}
\alias{checkInputNull}
\title{checkInputNull}
\usage{
checkInputNull(input_env, input_id)
}
\arguments{
\item{input_env}{shiny input object}

\item{input_id}{character string for inputId}
}
\value{
shinyFeedback object
}
\description{
checkInputNull
}
