% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checkInputNumeric.R
\name{checkInputNumeric}
\alias{checkInputNumeric}
\title{checkInputNumeric}
\usage{
checkInputNumeric(input_env, input_id, input_data)
}
\arguments{
\item{input_env}{shiny input object}

\item{input_id}{character string for inputId}

\item{input_data}{reactive shiny object representing data}
}
\value{
shinyFeedback object
}
\description{
checkInputNumeric
}
