% *=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=* 
% ** Copyright UCAR (c) 1992 - 2004 
% ** University Corporation for Atmospheric Research(UCAR) 
% ** National Center for Atmospheric Research(NCAR) 
% ** Research Applications Program(RAP) 
% ** P.O.Box 3000, Boulder, Colorado, 80307-3000, USA 
% ** 2004/9/1 14:13:29 
% *=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=* 
\name{disc.dat}
\alias{disc.dat}
\title{ Discrimination plot dataset.}
\usage{data(disc.dat)}
\docType{data}
\description{
This dataset is used to illustrate the \code{\link{discrimination.plot}}
function.

}
\keyword{datasets}
% docclass is data
% Converted by Sd2Rd version 1.21.
