% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Comparison_fct_fitModelComp.R
\name{fitModelComp}
\alias{fitModelComp}
\title{fitModelComp}
\usage{
fitModelComp(
  data,
  method,
  settings = NULL,
  value_x1,
  value_x2 = NULL,
  value_y1,
  value_y2 = NULL,
  ci_interval
)
}
\arguments{
\item{data}{data.frame input}

\item{method}{character string denoting modelling method}

\item{settings}{list of settings for "Deming" method only}

\item{value_x1}{character string denoting column containing method 1 values}

\item{value_x2}{character string denoting column containing method 1 duplicate values}

\item{value_y1}{character string denoting column containing method 2 values}

\item{value_y2}{character string denoting column containing method 2 duplicate values}

\item{ci_interval}{numeric CI interval}
}
\value{
mcr or brms model object
}
\description{
fitModelComp
}
