% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Diagnostic_fct_fitModelDiag.R
\name{fitModelDiag}
\alias{fitModelDiag}
\title{fitModelDiag}
\usage{
fitModelDiag(data, col_value, col_label, positive, ci_interval)
}
\arguments{
\item{data}{data.frame input}

\item{col_value}{character string representing column of values}

\item{col_label}{character string representing column of labels}

\item{positive}{character string representing a "positive" case}

\item{ci_interval}{numeric CI interval}
}
\value{
brms model object
}
\description{
fitModelDiag
}
