% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Trueness_fct_fitModelEQA.R
\name{fitModelEQA}
\alias{fitModelEQA}
\title{fitModelEQA}
\usage{
fitModelEQA(
  data,
  col_value_1,
  col_value_2 = NULL,
  col_mean,
  col_var,
  col_n,
  var_option,
  method,
  settings = NULL
)
}
\arguments{
\item{data}{data.frame input}

\item{col_value_1}{character string denoting column containing measurement values}

\item{col_mean}{character string denoting column EQA mean values}

\item{col_var}{character string denoting column EQA variance values}

\item{col_n}{character string denoting column EQA n values}

\item{var_option}{character string denoting variance type}

\item{col_value_2character}{string denoting column containing duplicate measurement values}
}
\value{
brms model object
}
\description{
fitModelEQA
}
