% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Imprecision_fct_fitModelPrec.R
\name{fitModelPrec}
\alias{fitModelPrec}
\title{fitModelPrec}
\usage{
fitModelPrec(data, x_var, y_var, qc_level, col_level)
}
\arguments{
\item{data}{data.frame input}

\item{x_var}{character string denoting column containing x values}

\item{y_var}{character string denoting column containing y values}

\item{qc_level}{character string/numeric denoting QC level}

\item{col_level}{character string denoting colour for output}
}
\value{
brms model object
}
\description{
fitModelPrec
}
