% *=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=* 
% ** Copyright UCAR (c) 1992 - 2004 
% ** University Corporation for Atmospheric Research(UCAR) 
% ** National Center for Atmospheric Research(NCAR) 
% ** Research Applications Program(RAP) 
% ** P.O.Box 3000, Boulder, Colorado, 80307-3000, USA 
% ** 2004/9/1 14:13:29 
% *=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=* 
\name{forecast.dat}
\alias{forecast.dat}
\title{
Forecast forecast dataset.
}
\description{
This spatial data set is an example used by the \code{\link{int.scale.verify}}
function.  It consists of the Nimrod forecasting system and was provided
by Barbara Casati.
}
\keyword{datasets}
% docclass is data
% Converted by Sd2Rd version 1.21.
