% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Reference_utils_getBayesFactor.R
\name{getBayesFactor}
\alias{getBayesFactor}
\title{getBayesFactor}
\usage{
getBayesFactor(model, prior_location)
}
\arguments{
\item{model}{brms fit from fitModelRef()}
}
\description{
getBayesFactor
}
