% *=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=* 
% ** Copyright UCAR (c) 1992 - 2004 
% ** University Corporation for Atmospheric Research(UCAR) 
% ** National Center for Atmospheric Research(NCAR) 
% ** Research Applications Program(RAP) 
% ** P.O.Box 3000, Boulder, Colorado, 80307-3000, USA 
% ** 2004/9/1 14:13:29 
% *=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=* 
\name{lookup}
\alias{lookup}
\title{ Lookup table for crps.circ. }
\description{ Saved lookup data to expedite integration in the crps.circ
  function. }
\seealso{\code{\link{crps.circ}} }


\keyword{datasets}
% docclass is data
% Converted by Sd2Rd version 1.21.
