% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_modelChecks.R
\name{modelChecks}
\alias{modelChecks}
\title{fitModelPrec}
\usage{
modelChecks(model, model_type, ci_interval)
}
\arguments{
\item{model}{fitted brms or rstanarm model object}

\item{model_type}{character string denoting the type of model ("imprecision", "regression", or "reference")}

\item{ci_interval}{numeric interval width}
}
\value{
list containing ggplot objects
}
\description{
fitModelPrec
}
