% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Comparison_fct_plotComparison.R
\name{plotComparison}
\alias{plotComparison}
\title{plotComparison}
\usage{
plotComparison(
  data,
  method,
  settings = NULL,
  value_x1,
  value_x2 = NULL,
  value_y1,
  value_y2 = NULL,
  x_name,
  y_name,
  model,
  plot_height
)
}
\arguments{
\item{data}{data.frame input}

\item{method}{character string denoting method}

\item{settings}{list of settings for "Deming" method only}

\item{value_x1}{character string denoting column containing method 1 values}

\item{value_x2}{character string denoting column containing method 1 duplicate values}

\item{value_y1}{character string denoting column containing method 2 values}

\item{value_y2}{character string denoting column containing method 2 duplicate values}

\item{x_name}{character string denoting x-axis label}

\item{y_name}{character string denoting y-axis label}

\item{model}{mcr or brms model object}

\item{plot_height}{numeric plot height}
}
\value{
plotly object
}
\description{
plotComparison
}
