% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Diagnostic_fct_plotDCA.R
\name{plotDCA}
\alias{plotDCA}
\title{plotDCA}
\usage{
plotDCA(data, model, col_value, col_label, positive, ci_interval, plot_dim)
}
\arguments{
\item{data}{data.frame input}

\item{model}{brms model object}

\item{col_label}{character string representing column of labels}

\item{positive}{character string representing a "positive" case}

\item{ci_interval}{numeric CI interval}

\item{plot_dim}{clientData from session}

\item{col_val}{ue character string representing column of values}
}
\value{
plotly object
}
\description{
plotDCA
}
