% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Diagnostic_fct_plotDiag.R
\name{plotDiag}
\alias{plotDiag}
\title{plotDiag}
\usage{
plotDiag(data, model, col_value, col_label, positive, ci_interval, plot_dim)
}
\arguments{
\item{data}{data.frame input}

\item{model}{brms model object}

\item{col_value}{character string representing column of values}

\item{col_label}{character string representing column of labels}

\item{positive}{character string representing a "positive" case}

\item{ci_interval}{numeric CI interval}

\item{plot_height}{clientData from session}
}
\value{
plotly object
}
\description{
plotDiag
}
