% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Trueness_fct_plotEQA.R
\name{plotEQA}
\alias{plotEQA}
\title{plotEQA}
\usage{
plotEQA(
  data,
  option_var,
  col_mean,
  col_var,
  col_n,
  col_value_1,
  col_value_2 = NULL,
  model,
  method,
  plot_height
)
}
\arguments{
\item{data}{data.frame input}

\item{option_var}{character string denoting variance type}

\item{col_mean}{character string denoting column containing mean EQA values}

\item{col_var}{character string denoting column containing EQA variance values}

\item{col_n}{character string denoting column containing EQA n values}

\item{col_value_1}{character string denoting column containing measurement values}

\item{col_value_2}{character string denoting column containing duplicate measurement values}

\item{model}{brms model object}

\item{plot_height}{numeric plot height}
}
\value{
plotly object
}
\description{
plotEQA
}
