% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Imprecision_fct_plotPrecData.R
\name{plotPrecData}
\alias{plotPrecData}
\title{plotPrecData}
\usage{
plotPrecData(data, x_var, y_var, qc_level, col_level = NULL, analyte, plot_dim)
}
\arguments{
\item{data}{data.frame input}

\item{x_var}{character string denoting column containing x values}

\item{y_var}{character string denoting column containing y values}

\item{qc_level}{numeric/character string denoting QC level}

\item{col_level}{numeric/character string denoting column containing QC levels}

\item{analyte}{character string denoting analyte name}

\item{plot_dim}{clientData from session}
}
\value{
plotly object
}
\description{
plotPrecData
}
