% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Reference_fct_plotRef.R
\name{plotRef}
\alias{plotRef}
\title{plotRef}
\usage{
plotRef(data, model, col_value_1, col_value_2 = NULL, bayes_factor)
}
\arguments{
\item{data}{data.frame containing input_file data}

\item{model}{brms fit from fitModelRef()}

\item{col_value_1}{character string containing column name}

\item{col_value_2}{character string containing column name}

\item{bayes_factor}{bayesfactor object returned from getBayesFactor()}
}
\description{
plotRef
}
