% *=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=* 
% ** Copyright UCAR (c) 1992 - 2004 
% ** University Corporation for Atmospheric Research(UCAR) 
% ** National Center for Atmospheric Research(NCAR) 
% ** Research Applications Program(RAP) 
% ** P.O.Box 3000, Boulder, Colorado, 80307-3000, USA 
% ** 2004/9/1 14:13:29 
% *=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=* 
\name{pop}
\docType{data}
\alias{pop}
\title{ Probability of precipitation (pop) data. }
\usage{data(pop)}
\description{
These datasets are used to illustrate several functions including
\code{\link{value}} and \code{\link{roc.plot}}.

These forecasts are summaries of 24-hour  probability of precipitation
forecasts that were made by the Finnish Meteorological Institute (FMI) during 2003, for daily
precipitation in the city of Tampere in south central Finland.  Light
precipitation is considered rainfall greater than .2 mm.  Rainfall
accumulation is considered values greater than 4.4 mm.  Rows of data
either missing forecasts or observations have been removed.

This data has been kindly provided by Dr. Pertti Nurmi of the Finnish
Meteorological Institute. 
\url{http://www.cawcr.gov.au/projects/verification/POP3/POP3.html}
}
\keyword{datasets}
% docclass is data
% Converted by Sd2Rd version 1.21.
