\name{prob.frcs.dat}
\docType{data}
\alias{prob.frcs.dat}
\usage{data(prob.frcs.dat) }
\title{ Probablisitic  Forecast Dataset. }
\description{
This data set is used as an example of data used by the \code{roc.plot}
function.  The first column contains a probablisitic forecast for
aviation icing.  The second column contains a logical variable
indicating whether or not icing was observed.  
}
\references{
  \emph{PROBABILITY FORECASTS OF IN-FLIGHT ICING
  CONDITIONS} Barbara G. Brown, Ben C. Bernstein, Frank McDonough and
  Tiffany A. O. Bernstein, 8th Conference on Aviation, Range, and
  Aerospace Meteorology, Dallas TX, 10-15 January 1999.
  }

\keyword{datasets}
% docclass is data
% Converted by Sd2Rd version 1.21.
