% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_readFile.R
\name{readFile}
\alias{readFile}
\title{readFile}
\usage{
readFile(file, headings)
}
\arguments{
\item{file}{character string fileInput}

\item{headings}{boolean denoting whether to use headings}
}
\value{
data.frame
}
\description{
readFile
}
