% *=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=* 
% ** Copyright UCAR (c) 1992 - 2004 
% ** University Corporation for Atmospheric Research(UCAR) 
% ** National Center for Atmospheric Research(NCAR) 
% ** Research Applications Program(RAP) 
% ** P.O.Box 3000, Boulder, Colorado, 80307-3000, USA 
% ** 2004/1/7 11:31:8 
% *=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=* 
\name{verify-internal}
\alias{table.stats}
\alias{summary.prob.bin}
\alias{summary.bin.bin}
\alias{summary.cont.cont}
\alias{summary.norm.dist.cont}
\alias{summary.cat.cat}
\alias{ranked.hist}
\alias{plot.prob.bin}
\alias{plot.cont.cont}
\alias{roc}
\alias{zzz}
\title{
  Verification internal and secondary functions
}
\description{
 Listed below are either internal functions or secondary functions.
 }
 \usage{
 table.stats(obs, pred)
 summary.prob.bin(object, \dots)
 summary.bin.bin(object, \dots)
 summary.cont.cont(object, \dots)
 summary.norm.dist.cont(object, \dots)
 summary.cat.cat(object, \dots)
 ranked.hist(frcst, nbins = 10, titl = NULL) 
 plot.prob.bin(x, \dots)
 plot.cont.cont(x, \dots)
 roc(x, pred, thres, binormal, \dots)
 zzz
}
\keyword{internal}
% docclass is function
% Converted by Sd2Rd version 1.21.
