% *=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=* 
% ** Copyright UCAR (c) 1992 - 2004 
% ** University Corporation for Atmospheric Research(UCAR) 
% ** National Center for Atmospheric Research(NCAR) 
% ** Research Applications Program(RAP) 
% ** P.O.Box 3000, Boulder, Colorado, 80307-3000, USA 
% ** 2004/9/1 14:13:29 
% *=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=* 
\name{precip.ensemble}
\alias{precip.ensemble}
\title{
An ensemble of precipitation forecasts 
}
\description{
  This is an example of an ensemble of precipitation forecasts.  The
  data set contains forecast for 517 days (3 monsoon seasons) at lead
  times of 1 to 10 days.  Observations and forecasts are in millimeters.
}
\keyword{datasets}
% docclass is data
% Converted by Sd2Rd version 1.21.
